# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: rechvix API
  version: "0.1.0"
  description: >
    Multi-company, multi-branch, multi-warehouse billing, inventory,
    accounting, and India-GST tax-management API. This spec is generated
    by reading `internal/modules/*/httpapi/handlers.go` directly — every
    path, method, and request/response shape below is verified against
    the actual Go handler code, not inferred from `docs/architecture.md`'s
    design intent. See `docs/api.md` for narrative conventions (auth,
    error envelope, money shape, pagination status) that apply across
    every endpoint below rather than being repeated per-operation.

    Status: complete — all 124 real routes across all 15 HTTP modules are
    present and typed, mechanically verified against a grep of every
    r.Get/Post/Put/Delete/Patch(...) registration in the codebase (exact
    match, see the git history for this file). Kept current as endpoints
    ship (docs/TODO.md's cross-cutting checklist item) — re-run that same
    grep diff after adding or changing a route, don't just trust this
    note.
servers:
  - url: /api/v1
    description: Same origin as the web app; also reachable directly on the configured HTTP_PORT.
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /logistics/vehicles:
    get:
      tags: [Logistics]
      summary: List vehicles.
      parameters: [{ name: active_only, in: query, schema: { type: boolean, default: true } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { vehicles: { type: array, items: { $ref: "#/components/schemas/Vehicle" } } } } } } }
    post:
      tags: [Logistics]
      summary: Create a vehicle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [registration_number, vehicle_type, default_transport_mode]
              properties:
                registration_number: { type: string }
                nickname: { type: string }
                vehicle_type: { type: string }
                default_transport_mode: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Vehicle" } } } }
  /logistics/vehicles/{id}/deactivate:
    post:
      tags: [Logistics]
      summary: Deactivate a vehicle.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Deactivated, no content. }
  /logistics/transporters:
    get:
      tags: [Logistics]
      summary: List transporters.
      parameters: [{ name: active_only, in: query, schema: { type: boolean, default: true } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { transporters: { type: array, items: { $ref: "#/components/schemas/Transporter" } } } } } } }
    post:
      tags: [Logistics]
      summary: Create a transporter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, default_transport_mode]
              properties:
                name: { type: string }
                transporter_id: { type: string }
                gstin: { type: string }
                phone: { type: string }
                address: { type: string }
                default_transport_mode: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Transporter" } } } }
  /logistics/transporters/{id}/deactivate:
    post:
      tags: [Logistics]
      summary: Deactivate a transporter.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Deactivated, no content. }

  # ============================================================
  # NOTIFICATIONS — share links (signed, expiring, revocable) and the
  # queued send path (Email/SMS/WhatsApp provider interfaces, mock only
  # — docs/TODO.md Stage 9). /share/{token} is the one deliberately
  # UNAUTHENTICATED endpoint in this entire API (brief §21 — a share
  # link's recipient has no session or API key by definition).
  # Verified against internal/modules/notifications/httpapi/handlers.go.
  # ============================================================
  /share-links:
    post:
      tags: [Notifications]
      summary: Create a signed, expiring share link for a document. The token is shown exactly once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_type, document_id]
              properties:
                document_type: { type: string }
                document_id: { type: string, format: uuid }
      responses:
        "201": { description: Created, content: { application/json: { schema: { type: object, properties: { token: { type: string } } } } } }
  /share-links/{id}:
    delete:
      tags: [Notifications]
      summary: Revoke a share link immediately.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { status: { type: string, example: revoked } } } } } }
  /share/{token}:
    get:
      tags: [Notifications]
      summary: >
        Redeem a share link — deliberately unauthenticated, the only such
        endpoint in this API. Proves the token is valid and reports what
        it grants access to; does not yet stream the shared document's
        actual PDF bytes (a documented follow-up — every existing print/
        report path requires a real permissions.Principal, which an
        anonymous recipient doesn't have by definition).
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_type: { type: string }
                  document_id: { type: string, format: uuid }
        "404":
          description: LINK_INVALID — invalid, expired, or revoked.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /notifications/send:
    post:
      tags: [Notifications]
      summary: Queue a notification (email/SMS/WhatsApp) for delivery. No real provider adapter is wired yet — mock only (docs/TODO.md Stage 9).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, recipient, document_type, document_id]
              properties:
                channel: { type: string, enum: [EMAIL, SMS, WHATSAPP] }
                recipient: { type: string }
                document_type: { type: string }
                document_id: { type: string, format: uuid }
                subject: { type: string }
                body_html: { type: string }
                template_name: { type: string, description: WhatsApp requires a pre-approved template (Business Platform shape), never freeform. }
                template_args: { type: array, items: { type: string } }
      responses:
        "202": { description: Queued., content: { application/json: { schema: { type: object, properties: { status: { type: string, example: queued } } } } } }

  # ============================================================
  # WEBHOOKS — endpoint registration only; delivery itself is outbound,
  # not a callable endpoint. Verified against
  # internal/modules/webhooks/httpapi/handlers.go.
  # ============================================================
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bp_session
      description: Set by POST /auth/login. Default cookie name; configurable via SESSION_COOKIE_NAME.
    apiKey:
      type: http
      scheme: bearer
      description: >
        Authorization: Bearer <key>. Effective permissions are the
        intersection of the key's declared scopes and the owning user's
        real RBAC grants — see docs/api.md.
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: SESSION_INVALID }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
...
    Vehicle:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        RegistrationNumber: { type: string }
        Nickname: { type: string }
        VehicleType: { type: string }
        DefaultTransportMode: { type: string }
        Active: { type: boolean }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    Transporter:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        Name: { type: string }
        TransporterID: { type: string }
        GSTIN: { type: string }
        Phone: { type: string }
        Address: { type: string }
        DefaultTransportMode: { type: string }
        Active: { type: boolean }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }